@@ -9,16 +9,23 @@
     "url": "https://github.com/kohlmeier/ssrs"
   },
   "devDependencies": {
-      "jsx-loader": "^0.11.0",
-      "es6-loader": "^0.2.0",
-      "react-tools": "^0.12.1",
-      "webpack": "^1.4.3",
-      "webpack-dev-server": "^1.6.5"
+    "babel-core": "^6.5.1",
+    "babel-loader": "^6.2.2",
+    "babel-plugin-react-transform": "^2.0.0",
+    "babel-preset-es2015": "^6.5.0",
+    "babel-preset-react": "^6.5.0",
+    "express": "^4.13.4",
+    "react-transform-catch-errors": "^1.0.2",
+    "react-transform-hmr": "^1.0.2",
+    "redbox-react": "^1.2.2",
+    "webpack": "^1.12.13",
+    "webpack-dev-middleware": "^1.5.1",
+    "webpack-hot-middleware": "^2.6.4"
   },
   "dependencies": {
-    "reactify": "~0.14.0",
-    "react-nested-router": "~0.2.1",
-    "react-tools": "~0.11.0",
-    "watchify": "~0.10.2"
+    "history": "^1.13.1",
+    "react": "^0.14.7",
+    "react-dom": "^0.14.7",
+    "react-router": "^1.0.3"
   }
 }
