@@ -1,10 +1,9 @@
-/** @jsx React.DOM */
 /*
  * Interface for feed mode
  */
-var React = require('react');
-var Link = require('react-nested-router').Link;
-var gravatar = require('./gravatar.js');
+import React from 'react';
+import { Link } from 'react-router';
+import gravatar from './gravatar.js';
 
 var UserHeader = React.createClass({
     getInitialState: function() {
